Fix findErrorPolicy cannot ensure result not null
Currently the findErrorPolicy do not guarantee the result is not null (if something wrong in defaultiwlanerrorconfig). With this fix we can ensure it return NonNull ErrorPolicy, and we can remove all redundant null policy checking in ErrorInfo

Code quality improvement only, this 'bug' does not causing any issue outside the code.

Bug: 307486851
Test: atest
Change-Id: I552a76b79a0a48c2277bfa4584aaf21a44f89f14