Sever dependency between IwlanDataService and EpdgSelector
Problem: Presently, IwlanDataService retains an instance of EpdgSelector and an instance of EpdgTunnelManager. EpdgTunnelManager additionally holds an instance of EpdgSelector. In fact, EpdgSelector should not be directly handled by IwlanDataService.

Solution:
- Sever the dependency between IwlanDataService and EpdgSelector. Transfer prefetch handling within EpdgTunnelManager.
- Initialize EpdgTunnelManager within IwlanDataService's constructor in
  order to initialize EpdgSelector early.

Benefit: This action enhances the readability and maintainability of the code. Additionally, it simplifies future modifications to the code, such as the removal of the singleton pattern.

Bug: 356493171
Flag: EXEMPT refactor
Test: atest IwlanTests
Test: On-device tests
Change-Id: I276d3c9da7e95fd593377ae3d916138f7cc5d43b