Fix metrics cleared before tunnel fully opened
Before this fix, after a tunnel is closed, the tunnel metrics are not immediately cleaned up. If a setup data call request on the same APN type is received before the tunnel metrics are cleared, the metrics will be cleaned up, and the metrics for the new tunnel will not be found.
The solution is to synchronize the lifecycle of tunnel metrics with tunnel state changes by utilizing a single callback for both tunnel open and tunnel close events.

Bug: 357629949
Flag: EXEMPT refactor
Test: atest IwlanTests
Test: On-device tests
Change-Id: I1e9ccc611c2c2c018ccc2032157479247cfc26d8